projetos/banco-dados/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
